[refac]: moved class SSRQConfig to own file
